requirements.txt
# ========== AI & LLM ==========
openai==1.3.0
anthropic==0.7.0

# ========== HTTP & Network ==========
requests==2.31.0
aiohttp==3.9.0
urllib3==2.0.7

# ========== Web Framework ==========
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.0

# ========== CLI & UI ==========
rich==13.7.0
click==8.1.7
colorama==0.4.6
prompt-toolkit==3.0.41

# ========== Database ==========
sqlalchemy==2.0.23

# ========== Environment ==========
python-dotenv==1.0.0

# ========== Testing ==========
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# ========== Code Quality ==========
black==23.12.0
flake8==6.1.0
pylint==3.0.3

# ========== Documentation ==========
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# ========== Security ==========
cryptography==41.0.7
